acceptance: moving the older helpers into a deprecated package to help us move
